requirements.txt
django
psycopg2-binary
django-leaflet
django-leaflet-storage
psycopg2-binary
django-import-export
djangorestframework
pillow
drf-yasg
django-cors-headers
graphene-django
django-filter
paramiko
python-telegram-bot
celery
redis
requests
django-cron
django-autocomplete-light
django-crontab
openpyxl
djangorestframework-gis
geopy
reportlab
aiogram==2.25.1
asgiref>=3.3.2,<4
watchdog==2.1.6
PyYAML==6.0
apscheduler
django-colorful
django-revproxy
prefect
geopandas
shapely
pysal
esda
djangorestframework-simplejwt